requirements.txt
flask==3.1.0
gunicorn==25.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
apscheduler==3.11.0
cloudscraper==1.2.71
pywebpush==2.0.4
py-vapid==1.9.2
pytz==2025.1
cryptography==44.0.2
requests==2.32.3
email_validator==2.2.0
